Refactor auth service, remove user table, and update assets
Moved auth-service from lib to services and updated imports accordingly. Removed the unused UserTable component. Renamed logo asset and updated its usage in the login page. Improved error logging in chat-table, simplified asset-type-summary performance hooks, and made minor code cleanups in performance utilities and dashboard widgets.
